set Dessert and MainCourse test in separated files
