+++ src/main/resources/application-local.properties
@@ -16,7 +16,7 @@ spring.datasource.password=
 #spring.jpa.properties.hibernate.format_sql=true
 
 # DDL OPTIONS: create-drop, create, update, none, validate
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=none
 # create ? Hibernate first drops existing tables and then creates new tables.
 # update ? The object model created based on the mappings
 # (annotations or XML) is compared with the existing schema,
